+++ changelog.md
@@ -9,6 +9,7 @@ Changelog
 + Add `text.format.toString`
 + Add `Text.Format.defaultFont`, `Text.Format.defaultHorizAlign`, `Text.Format.defaultVertAlign`
 + Modify `Text.Format` constructor default values
++ Fix incorrect values set in `rac.Text.Format.topLeft` and `rac.Text.Format.topRight`
 
 
 1.1.0 - 2021 October 13th