+++ changelog.md
@@ -11,6 +11,7 @@ Changelog
 + Modify `Text.Format` constructor defaults and parameter order
 + Modify `Text.Format` to allow nullable `size` and `font`
 + Fix incorrect values set in `rac.Text.Format.topLeft` and `rac.Text.Format.topRight`
++ Modify `point.text` to allow an optional `format` parameter
 
 
 1.1.0 - 2021 October 13th